Merge conflict fix WIP
Merging the coin PR seemed to break a bunch of stuff. The player prefab wasn't usable anymore so I recreated it. Now there seems to be a weird bug where the player stalls for way too long, and I can no longer get the CoinPickup script to reference the UpdateCount Script. I will find a new way to organize the coin code